Initial success with exporting to ONNX
This code exports the model as a static graph. The graph can be inspected with https://netron.app/

I hardcoded the T value to 767 in order to avoid the following error:
RuntimeError: Unsupported: ONNX export of convolution for kernel of unknown shape.

I don't understand the architecture of RWKV-3, I'm just poking around. I think this hack has
a detrimental effect if your context is not full.