Exported RWKV_RNN() works now
I don't pre-init it the way it is done in run.py and that definitely affects the output: it feels like it tries
to continue a fragment of a book instead of a news article. Either way, the output is coherent.

I made a mistake of feeding it a [1][768] input when it needed [768]. That took a while to track down.